% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genres_by_age_2018}
\alias{genres_by_age_2018}
\title{Favorite music genres among consumers in the United States as of July 2018, by age group}
\format{A data frame with 84 rows and 3 variables:
\describe{
\item{genre}{The genre of music}
\item{agr_range}{The agr range associated with the responce}
\item{percent_responce}{The percentage responce from the given age range for the given genre}
}}
\source{
\url{https://www.statista.com/statistics/253915/favorite-music-genres-in-the-us/}
}
\usage{
genres_by_age_2018
}
\description{
"The statistic provides data on favorite music genres among consumers in the United States
as of July 2018, sorted by age group. According to the source, 52 percent of respondents
aged 16 to 19 years old stated that pop music was their favorite music genre, compared to
19 percent of respondents aged 65 or above."
https://www.statista.com/statistics/253915/favorite-music-genres-in-the-us/
}
\details{
"In 2012, country music topped the list; 27.6 percent of respondents picked it among their
three favorite genres. A year earlier, the result was one percent lower, which allowed
classic rock to take the lead. The figures show, however, the genre’s popularity across
the United States is unshakeable and it has also been spreading abroad. This could be
demonstrated by the international success of (among others) Shania Twain or the second
place the Dutch country duo “The Common Linnets” received in the Eurovision Song Contest
in 2014, singing “Calm after the storm.”

The genre is also widely popular among American teenagers, earning the second place and
15.3 percent of votes in a survey in August 2012. The first place and more than 18 percent
of votes was awarded to pop music, rock scored 13.1 percent and landed in fourth place.
Interestingly, Christian music made it to top five with nine percent of votes. The younger
generation is also widely represented among country music performers with such prominent
names as Taylor Swift (born in 1989), who was the highest paid musician in 2015, and Hunter
Hayes (born in 1991). Country music is also able to attract crowds (and large sums of money)
to live performances. Luke Bryan’s tour was the most successful tour in North America in 2016
based on ticket sales as almost 1.43 million tickets were sold for his shows. Fellow country
singer, Garth Brooks, came second on the list, selling 1.4 million tickets for his tour in
North America in 2016." https://www.statista.com/statistics/253915/favorite-music-genres-in-the-us/

"This question was phrased by the source as follows: "Which of the following types of music do you generally like?""
https://www.statista.com/statistics/253915/favorite-music-genres-in-the-us/

Region: United States
Survey time period: July 2018
Number of respondents: 3,000 respondents
Age group: 16 years and older
Method of interview: Online survey
}
\keyword{datasets}
