% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{listening_frequency_by_age_2017}
\alias{listening_frequency_by_age_2017}
\title{Music listening frequency in the United States as of August 2017, by age}
\format{A data frame with 24 rows and 3 variables:
\describe{
\item{Generation}{The generational age range for the responce}
\item{Frequency}{The frequency the generation listens to music}
\item{percent}{The percentage responce from the given Generation for the given Frequency}
}}
\source{
\url{https://www.statista.com/statistics/749666/music-listening-habits-age-usa/}
}
\usage{
listening_frequency_by_age_2017
}
\description{
"As of August 2017, 84 percent of Gen Z adults reported listening to music every day, and
the majority of Millennials as well as their older peers said the same. Daily music
listening was far more commonplace among all respondents than weekly or monthly listening,
with very few adults stating that they never listened to music (or did so on a monthly or
less frequent basis)." https://www.statista.com/statistics/749666/music-listening-habits-age-usa/
}
\details{
"Despite the many genres and formats available (which in theory cater to every kind of music
fan out there) music consumption does change with age. For example, the share of adults aged
18 to 29 years old who are avid music fans is far higher than for other age groups, and
almost triple that of adults aged over 65. Similarly, rock ‘n’ roll and country/western music
is far more popular among older generations, who have little enthusiasm for hip-hop and rap
music, which are much more preferable among 16 to 19-year-olds.

It is not just a passion for music or a specific genre which changes, but also the way in
which music is consumed. For example, studies have shown that the popularity of music
streaming sources generally dwindles with age, and platforms like YouTube reach more adults
aged under 25 years than older generations, whereas Baby Boomers are more willing to pay for
music on vinyl records than Millennials or Gen Z respondents."
https://www.statista.com/statistics/749666/music-listening-habits-age-usa/

Region: United States
Survey time period: August 22, 2017
Number of respondents: 1,900 respondents
Age group: 18 years and older
Method of interview: Online survey
}
\keyword{datasets}
