% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{racially_diverse_music_2018}
\alias{racially_diverse_music_2018}
\title{Public opinion on the racial diversity of selected music genres in the United States as of May 2018}
\format{A data frame with 21 rows and 2 variables:
\describe{
\item{genre}{The genre of music}
\item{percent_responce}{The percentage responce for the given genre}
}}
\source{
\url{https://www.statista.com/statistics/864622/music-genre-diversity/}
}
\usage{
racially_diverse_music_2018
}
\description{
"This statistic shows the public opinion on the racial diversity of selected music genres in the
United States as of May 2018, by age. During the survey, 25 percent of respondents stated that
they considered rap/hip-hop to be the most racially diverse music genre."
https://www.statista.com/statistics/864622/music-genre-diversity/
}
\details{
"This question was phrased by the source as follows: "In which, if any, of the following
genres do you find artists to be more racially/ethnically diverse than other genres?""
https://www.statista.com/statistics/864622/music-genre-diversity/

Region: United States
Survey time period: May 10 - 11, 2018
Number of respondents: 1,108 respondents
Age group: 18 years and older
Method of interview: Online survey
}
\keyword{datasets}
